end of  part 4

+++ part4/blog/controllers/blogs.js
@@ -4,19 +4,13 @@ const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
 blogRouter.get('/', async (request, response) => {
-    const resp = await Blog.find({}).populate('user')
+    const resp = await Blog.find({}).populate('user', 'username name id')
     return response.json(resp)
 })
   
 blogRouter.post('/', async (request, response) => {
     const blog = new Blog(request.body)
-
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token invalid' })
-    }
-    const user = await User.findById(decodedToken.id)
-    blog.user = user
+    blog.user = request.user
 
     if (blog.likes === undefined) {
         blog.likes = 0 
@@ -25,23 +19,29 @@ blogRouter.post('/', async (request, response) => {
     if (blog.title === undefined || blog.url === undefined) {
         return response.status(400).end()
     }    
-    const resp = await blog.save()
-
+    const resp = await blog.save() 
+    let user = request.user
     user.blogs = user.blogs.concat(resp._id)
-    await user.save()
-
+    const u = await user.save()
     return response.status(201).json(resp)
 })
 
 blogRouter.delete('/:id', async (request, response) => {
     try {
-        const deletedObject = await Blog.findByIdAndDelete(request.params.id)
+        const deletedObject = await Blog.findById(request.params.id)
+
+        if (deletedObject.user.toString() !== request.user._id.toString()) {
+            return response.status(401).json({ error: 'token invalid' })
+        }
         
         if (!deletedObject) {
             return response.status(404).end()
         }
+
+        await Blog.findByIdAndDelete(request.params.id)
         return response.status(204).end()
     } catch (err) {
+        console.log(err)
         return response.status(500).end()
     }
 })
